+++ src/main/java/com/sparksInTheStep/webBoard/comment/presentation/dto/CommentResponse.java
@@ -1,6 +1,6 @@
 package com.sparksInTheStep.webBoard.comment.presentation.dto;
 
-import com.sparksInTheStep.webBoard.auth.presentation.dto.MemberResponse;
+import com.sparksInTheStep.webBoard.member.presentation.dto.MemberResponse;
 import com.sparksInTheStep.webBoard.comment.application.dto.CommentInfo;
 import com.sparksInTheStep.webBoard.post.controller.dto.PostResponse;
