+++ src/main/java/com/sparksInTheStep/webBoard/global/filter/MemberArgumentResolver.java
@@ -1,7 +1,6 @@
 package com.sparksInTheStep.webBoard.global.filter;
 
-import com.sparksInTheStep.webBoard.auth.application.MemberService;
-import com.sparksInTheStep.webBoard.auth.application.dto.MemberInfo;
+import com.sparksInTheStep.webBoard.member.application.MemberService;
 import com.sparksInTheStep.webBoard.auth.token.JwtTokenProvider;
 import com.sparksInTheStep.webBoard.global.annotation.AuthorizedUser;
 import com.sparksInTheStep.webBoard.global.errorHandling.CustomException;
